+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/code-editor/common/common.ts
