+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/execution-result/ExecutionResult.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useCallback } from 'react';
-import { ITestRunType } from '../../../hooks/submissions/use-submissions';
+import { ITestRunType } from '../../../hooks/submissions/types';
 import concatClassNames from '../../../utils/class-names';
 import styles from './ExecutionResult.module.scss';
