+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/hooks/submissions/use-problem-submissions.tsx
@@ -6,9 +6,9 @@ import { ISubmissionDetails } from './types';
 import { useLoading } from '../use-loading';
 import { useHttp } from '../use-http';
 import { useUrls } from '../use-urls';
-import { DEFAULT_PROBLEM_RESULTS_TAKE_CONTESTS_PAGE } from '../../common/constants';
+import { DEFAULT_PROBLEM_RESULTS_TAKE_CONTESTS_PAGE } from '../../components/code-editor/common/constants';
 import { useCurrentContest } from '../use-current-contest';
-import { UrlType } from '../../common/common-types';
+import { UrlType } from '../../components/code-editor/common/common-types';
 
 interface IProblemSubmissionsContext {
     state: {
