+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/hooks/use-contest-categories.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useCallback, useContext, useEffect, useState } from 'react';
 import { isEmpty } from 'lodash';
 import { IHaveChildrenProps } from '../components/common/Props';
-import { IContestCategoryTreeType } from '../common/types';
+import { IContestCategoryTreeType } from '../components/code-editor/common/types';
 import { useHttp } from './use-http';
 import { useUrls } from './use-urls';
 import { useLoading } from './use-loading';
